SMall refactoring
